Modernizing; updating this gem; 100% test coverage
 — following Gem naming style guide for location of files and folders
 — created a new `Client` class, but made top level namespace backwards compatible.
 — Added a couple of new tests to bring coverage to 100%
 — Added Travis CI and CodeClimate
 — Added SimpleCov
 — replaced deprecated usages of `stub(blah)` with `expect(blah).to receive`
