+++ .gitignore
@@ -8,6 +8,7 @@
 /test/tmp/
 /test/version_tmp/
 /tmp/
+.idea/
 
 ## Specific to RubyMotion:
 .dat*
@@ -32,3 +33,4 @@ build/
 
 # unless supporting rvm < 1.11.0 or doing something fancy, ignore this:
 .rvmrc
+Gemfile.lock
