+++ .rspec
@@ -1,0 +1,3 @@
+--format progress
+--color
+--order random
