+++ .travis.yml
@@ -1,0 +1,18 @@
+env:
+  global:
+    - CC_TEST_REPORTER_ID=7fcb20b914b9f973172b8d4506b6d29093995c7107580cbc2f3e55ac24227f05
+sudo: false
+language: ruby
+cache: bundler
+rvm:
+  - 2.4.1
+  - 2.3.5
+before_install: gem install bundler -v 1.15.4
+before_script:
+  - curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
+  - chmod +x ./cc-test-reporter
+  - ./cc-test-reporter before-build
+script:
+  - bundle exec rspec
+after_script:
+  - ./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT
