+++ lib/dnsmadeeasy/rest/api/version.rb
@@ -1,0 +1,7 @@
+module DnsMadeEasy
+  module Rest
+    module Api
+      VERSION = '1.1.0'.freeze
+    end
+  end
+end
